Roll back unverified tree if an error occurs when scanning new entries
Previously, the unverified tree would be left as it was during the scan.
This resulted in those entries not being re-scanned (and so the error was
lost to the depths of time) and the log becoming wedged due to having a
pending STH but no new entries to fetch (this wedging error was fixed by
an earlier commit, however).